Adjusted image sizes & colours.
I doubled the backgrounds side-by-side so when scaled up, they fit much
better with the camera/device orientation. As a square image, we had to
lose a lot of height just in order to get enough width to keep the
backgrounds on the screen. This sizing was applied to all other
environment images as well - specifically for the foreground so we
wouldn't have so much clutter up front.

I also made sure there was a window in every 1st level background
(offices) so that they could all be made 16-bit and not lose colour
consistency by some not having alpha representation. If this has to be
reversed for any reason, use "truecolour" setting.